+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### Unreleased
   - Add codacy badge
+  - Initial version of the SEFT publisher service